programs/requirements.txt
pip
Flask
requests
cassandra-driver